@@ -130,6 +130,7 @@
     "@ory/kratos-client": "^0.10.1",
     "@radix-ui/react-dialog": "^1.0.3",
     "@radix-ui/react-dropdown-menu": "^2.0.4",
+    "@supabase/supabase-js": "^2.26.0",
     "@tailwindcss/forms": "^0.5.2",
     "@tailwindcss/typography": "^0.5.9",
     "@xelah/type-perf-html": "^1.0.1",
@@ -138,6 +139,7 @@
     "bible-reference-rcl": "1.1.0",
     "clsx": "1.1.1",
     "crypto-js": "^4.1.1",
+    "dotenv": "^16.3.1",
     "electron-is-dev": "^2.0.0",
     "electron-log": "4.4.7",
     "electron-next": "^3.1.5",
@@ -159,7 +161,7 @@
     "markdown-translatable": "1.3.0",
     "md5": "^2.3.0",
     "moment": "^2.29.3",
-    "next": "^13.0.0",
+    "next": "^13.4.10",
     "next-images": "^1.8.4",
     "next-on-netlify": "^3.0.1",
     "next-optimized-images": "^2.6.2",
