@@ -1,5 +1,7 @@
+/* eslint-disable no-nested-ternary */
 /* eslint-disable react/jsx-no-useless-fragment */
-import React from 'react';
+
+import { useContext, useState, useEffect } from 'react';
 import Login from './components/Login/Login';
 import AuthenticationContextProvider, { AuthenticationContext } from './components/Login/AuthenticationContextProvider';
 import { loadUsers } from './core/Login/handleJson';
@@ -12,11 +14,10 @@ import { getorPutAppLangage } from './core/projects/handleProfile';
 import i18n from './translations/i18n';
 
 const Home = () => {
-  const { states, action } = React.useContext(AuthenticationContext);
-  const [token, setToken] = React.useState();
-  const [user, setUser] = React.useState();
-
-  React.useEffect(() => {
+  const { states, action } = useContext(AuthenticationContext);
+  const [token, setToken] = useState();
+  const [user, setUser] = useState();
+  useEffect(() => {
     logger.debug('Home.js', 'Triggers loadUsers for the users list');
     loadUsers();
   }, []);
@@ -41,9 +42,9 @@ const Home = () => {
     }
   };
 
-  React.useEffect(() => {
+  useEffect(() => {
     validateUser();
-  // eslint-disable-next-line react-hooks/exhaustive-deps
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [states.accessToken]);
 
   return (
