@@ -78,9 +78,9 @@ const { t } = useTranslation();
   return (
     <div>
       { isLoading === false ? (
-        <>
+        <div>
           {
-            stories.map((story, index) => (
+            stories?.map((story, index) => (
               <div
                 key={story.id}
                 className={`flex gap-5 mb-5 items-center justify-center ${story.id === selectedStory && 'bg-light'}`}
@@ -123,7 +123,7 @@ const { t } = useTranslation();
               </div>
             ))
           }
-        </>
+        </div>
         ) : (
           <LoadingScreen />
         )}
