@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { ChevronUpDownIcon } from '@heroicons/react/24/solid';
 
 function CustomMultiComboBox({
-  selectedList, setSelectedList, customData, placeholder = 'Select Language', filterParams = 'name', multiSelect = false, dropArrow = false, showLangCode = { show: false, langkey: 'lc' },
+  selectedList, setSelectedList, customData, filterParams = 'name', multiSelect = false, dropArrow = false, showLangCode = { show: false, langkey: 'lc' },
 }) {
   let filteredData = [];
   const [query, setQuery] = useState('');
@@ -32,7 +32,7 @@ function CustomMultiComboBox({
                   className="w-full border-none py-2 pl-3 pr-10 text-sm leading-5 text-gray-900 focus:ring-0"
                   // displayValue={(language) => language?.ang}
                   displayValue={(selectedList) => `${selectedList.length > 0 ? `${selectedList[0][filterParams]}${multiSelect ? '... click for more' : '' }` : ''}`}
-                  placeholder={placeholder}
+                  placeholder="Select Language"
                   onFocus={() => !open && setIsActive(true)}
                   onBlur={() => setIsActive(false)}
                   onChange={(event) => setQuery(event.target.value)}
