@@ -5,10 +5,11 @@ import * as localforage from 'localforage';
 import { splitStringByLastOccurance } from '@/util/splitStringByLastMarker';
 import { isElectron } from '../../core/handleElectron';
 import * as logger from '../../logger';
-import saveProjectsMeta from '../../core/projects/saveProjetcsMeta';
+import { saveProjectsMeta, saveSupabaseProjectsMeta } from '../../core/projects/saveProjetcsMeta';
 import { environment } from '../../../environment';
 import staicLangJson from '../../lib/lang/langNames.json';
 import packageInfo from '../../../../package.json';
+import { newPath, supabaseStorage } from '../../../../supabase';
 
 const path = require('path');
 const advanceSettings = require('../../lib/AdvanceSettings.json');
@@ -94,7 +95,53 @@ const ProjectContextProvider = ({ children }) => {
     }
     fs.writeFileSync(file, JSON.stringify(json));
   };
-
+  const createWebSettingJson = (file) => {
+    logger.debug('ProjectContext.js', 'Loading data from AdvanceSetting.json file');
+    setCanonList(advanceSettings.canonSpecification);
+    setLicenseList((advanceSettings.copyright).push({
+      id: 'Other', title: 'Custom', licence: '', locked: false,
+    }));
+    // setLanguages([advanceSettings.languages]);
+    const json = {
+      version: environment.AG_USER_SETTING_VERSION,
+      history: {
+        copyright: [{
+          id: 'Other', title: 'Custom', licence: '', locked: false,
+        }],
+        languages: [],
+        textTranslation: {
+          canonSpecification: [{
+            id: 4, title: 'Other', currentScope: [], locked: false,
+          }],
+        },
+      },
+      appLanguage: 'en',
+      theme: 'light',
+      userWorkspaceLocation: '',
+      commonWorkspaceLocation: '',
+      resources: {
+        door43: {
+          translationNotes: [],
+          translationQuestions: [],
+          translationWords: [],
+          obsTranslationNotes: [],
+        },
+      },
+      sync: { services: { door43: [] } },
+    };
+    console.debug('ProjectContext.js', `Creating a ${environment.USER_SETTING_FILE} file`);
+    const { data } = supabaseStorage().list(file);
+    console.log('ProjectContext.js', { data });
+    if (data.length === 0) {
+      const { data: envSettings } = supabaseStorage().upload(file, JSON.stringify(json), {
+        cacheControl: '3600',
+        upsert: false,
+      });
+      if (envSettings) {
+        console.log('ProjectContext.js', { envSettings });
+      }
+    }
+  };
   const concatLanguages = async (json, staicLangJson) => {
     logger.debug('ProjectContext.js', 'In concat languages');
     const userlanguages = [];
@@ -168,6 +215,51 @@ const ProjectContextProvider = ({ children }) => {
       createSettingJson(fs, file);
     }
   };
+
+  const loadWebSettings = async () => {
+    let currentUser;
+    await localforage.getItem('userProfile').then((value) => {
+      currentUser = value.user.email;
+      setUsername(value.user.email);
+    });
+    if (!currentUser) {
+      console.error('ProjectContext.js', 'Unable to find current user');
+    }
+
+    const file = `${newPath}/${currentUser}/${environment.USER_SETTING_FILE}`;
+    const { data: agUserSettings, error } = await supabaseStorage().download(file);
+    if (error) {
+      console.error('ProjectContext.js', 'Failed to read the data from file');
+    }
+    const json = JSON.parse(await agUserSettings.text());
+    if (json.version === environment.AG_USER_SETTING_VERSION) {
+      if (json.history?.copyright) {
+        if (json.history?.copyright?.licence) {
+          setLicenseList((advanceSettings.copyright)
+            .concat(json.history?.copyright));
+        } else {
+          const newObj = (advanceSettings.copyright).filter((item) => item.Id !== 'Other');
+          newObj.push({
+            id: 'Other', title: 'Custom', licence: '', locked: false,
+          });
+          setLicenseList(newObj);
+        }
+      } else {
+        setLicenseList(advanceSettings.copyright);
+      }
+      setCanonList(json.history?.textTranslation.canonSpecification
+        ? (advanceSettings.canonSpecification)
+          .concat(json.history?.textTranslation.canonSpecification)
+        : advanceSettings.canonSpecification);
+      // concat static and custom languages
+      const langFilter = await concatLanguages(json, staicLangJson);
+      const filteredLang = langFilter.concatedLang.filter((lang) => lang?.ang.trim() !== '');
+      setLanguages([...filteredLang]);
+      setCustomLanguages(langFilter.userlanguages);
+    } else {
+      createWebSettingJson(file);
+    }
+  };
   // Json for storing advance settings
   const updateJson = async (currentSettings) => {
     logger.debug('ProjectContext.js', 'In updateJson');
@@ -221,6 +313,53 @@ const ProjectContextProvider = ({ children }) => {
       }
     }
   };
+
+  const updateWebJson = async (currentSettings) => {
+    let currentUser;
+    await localforage.getItem('userProfile').then((value) => {
+      currentUser = value.user.email;
+      setUsername(value.user.email);
+    });
+    const file = `${newPath}/${currentUser}/${environment.USER_SETTING_FILE}`;
+    const { data } = await supabaseStorage().download(file);
+    if (data) {
+      const json = JSON.parse(await data.text());
+      // eslint-disable-next-line no-nested-ternary
+      const currentSetting = (currentSettings === 'copyright' ? copyright
+        : (currentSettings === 'languages' ? {
+          title: language.ang,
+          id: language.id,
+          scriptDirection: language.ld,
+          langCode: language.lc,
+          custom: true,
+        }
+          : canonSpecification));
+      if (currentSettings === 'canonSpecification') {
+        (json.history?.textTranslation[currentSettings])?.push(currentSetting);
+      } else if (json.history[currentSettings]
+        && uniqueId(json.history[currentSettings], currentSetting.id)) {
+        (json.history[currentSettings]).forEach((setting) => {
+          if (setting.id === currentSetting.id) {
+            const keys = Object.keys(setting);
+            keys.forEach((key) => {
+              setting[key] = currentSetting[key];
+            });
+          }
+        });
+      } else {
+        // updating the canon or pushing new language
+        (json.history[currentSettings]).push(currentSetting);
+      }
+      json.version = environment.AG_USER_SETTING_VERSION;
+      json.sync.services.door43 = json?.sync?.services?.door43 ? json?.sync?.services?.door43 : [];
+      console.debug('ProjectContext.js', 'Upadting the settings in existing file');
+      await supabaseStorage().upload(file, JSON.stringify(json));
+      console.debug('ProjectContext.js', 'Loading new settings from file');
+      await loadWebSettings();
+    } else {
+      console.error('ProjectContext.js', 'Failed to read the data from file');
+    }
+  };
   // common functions for create projects
   const createProjectCommonUtils = async () => {
     logger.debug('ProjectContext.js', 'In createProject common utils');
@@ -231,18 +370,18 @@ const ProjectContextProvider = ({ children }) => {
           if (lang.id === language.id) {
             if (lang.ang !== language.ang
               || lang.ld !== language.ld || lang.lc !== language.lc) {
-              await updateJson('languages');
+              isElectron() ? await updateJson('languages') : await updateWebJson('languages');
             }
           }
         });
       } else {
         // add language to custom
-        await updateJson('languages');
+        isElectron() ? await updateJson('languages') : await updateWebJson('languages');
       }
     }
     // Update Custom licence into current list.
     if (copyright.title === 'Custom') {
-      updateJson('copyright');
+      isElectron() ? await updateJson('copyright') : await updateWebJson('copyright');
     } else {
       const myLicence = Array.isArray(licenceList) ? licenceList.find((item) => item.title === copyright.title) : [];
       // eslint-disable-next-line import/no-dynamic-require
@@ -257,7 +396,7 @@ const ProjectContextProvider = ({ children }) => {
     logger.debug('ProjectContext.js', 'In createProject Translation utils');
     // Update Custom canon into current list.
     if (canonSpecification.title === 'Other') {
-      await updateJson('canonSpecification');
+      isElectron() ? await updateJson('canonSpecification') : await updateWebJson('canonSpecification');
     }
   };
 
@@ -288,6 +427,31 @@ const ProjectContextProvider = ({ children }) => {
     return status;
   };
 
+  const createSupabaseProject = async (call, project, update, projectType) => {
+    createProjectCommonUtils();
+    // common props pass for all project type
+    const projectMetaObj = {
+      newProjectFields,
+      language,
+      copyright,
+      importedFiles,
+      call,
+      project,
+      update,
+      projectType,
+    };
+    if (projectType !== 'OBS') {
+      createProjectTranslationUtils();
+      const temp_obj = {
+        versificationScheme: versificationScheme.title,
+        canonSpecification,
+      };
+      Object.assign(projectMetaObj, temp_obj);
+    }
+    logger.debug('ProjectContext.js', 'Calling saveSupabaseProjectsMeta with required props');
+    const status = saveSupabaseProjectsMeta(projectMetaObj);
+    return status;
+  };
   const resetProjectStates = () => {
     const initialState = {
       language: '',
@@ -324,7 +488,31 @@ const ProjectContextProvider = ({ children }) => {
           });
         });
       });
+    } else {
+      loadWebSettings();
+      localforage.getItem('userProfile').then((value) => {
+        setUsername(value?.user?.email);
+      });
+      localforage.getItem('currentProject').then((projectName) => {
+        setSelectedProject(projectName);
+        // setProjectMeta in a var
+        localforage.getItem('projectmeta').then((projectMeta) => {
+          setSelectedProject(projectName);
+          // setProjectMeta in a var
+          projectMeta?.projects.forEach((meta) => {
+            const currentprojectId = Object.keys(meta.identification.primary[packageInfo.name])[0];
+            const currentprojectName = meta.identification.name.en;
+            splitStringByLastOccurance(projectName, '_').then((arr) => {
+              if (arr.length > 0 && arr[0].toLowerCase() === currentprojectName.toLowerCase()
+                && arr[1].toLowerCase() === currentprojectId.toLocaleLowerCase()) {
+                setSelectedProjectMeta(meta);
+              }
+            });
+          });
+        });
+      });
     }
+
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
@@ -359,6 +547,7 @@ const ProjectContextProvider = ({ children }) => {
       setSideTabTitle,
       setSelectedProject,
       createProject,
+      createSupabaseProject,
       setLanguage,
       setScrollLock,
       setUsername,
