@@ -1,18 +1,20 @@
 import localforage from 'localforage';
 // import { readRefMeta } from '../../../core/reference/readRefMeta';
+import { isElectron } from '@/core/handleElectron';
 import { readRefBurrito } from '../../../core/reference/readRefBurrito';
 import writeToFile from '../../../core/editor/writeToFile';
 import packageInfo from '../../../../../package.json';
+import { newPath } from '../../../../../supabase';
 // function to save to file.
 export const saveToFile = async (usfmText, bookCode) => {
   try {
     const userProfile = await localforage.getItem('userProfile');
-    const userName = userProfile?.username;
+    const userName = isElectron() ? userProfile?.username : userProfile?.user?.email;
     const projectName = await localforage.getItem('currentProject');
     const path = require('path');
     const newpath = localStorage.getItem('userPath');
     // const projectsDir = path.join(newpath, packageInfo.name, 'users', userName, 'projects', projectName);
-    const metaPath = path.join(newpath, packageInfo.name, 'users', userName, 'projects', projectName, 'metadata.json');
+    const metaPath = isElectron() ? path.join(newpath, packageInfo.name, 'users', userName, 'projects', projectName, 'metadata.json') : `${newPath}/${userName}/projects/${projectName}/metadata.json`;
     // const refs = await readRefMeta({ projectsDir })
     const metaData = JSON.parse(await readRefBurrito({ metaPath }));
     Object.entries(metaData.ingredients).forEach(async ([key, _ingredients]) => {
