@@ -3,6 +3,8 @@ import { splitStringByLastOccurance } from '@/util/splitStringByLastMarker';
 import * as logger from '../../logger';
 import { environment } from '../../../environment';
 import packageInfo from '../../../../package.json';
+import { newPath, supabaseStorage } from '../../../../supabase';
+import { isElectron } from '../handleElectron';
 
 export const updateAgSettings = async (username, projectName, data) => {
   logger.debug('updateAgSettings.js', 'In updateAgSettings');
@@ -18,13 +20,35 @@ export const updateAgSettings = async (username, projectName, data) => {
     if (!setting.sync && !setting.sync?.services) {
       setting.sync = { services: { door43: [] } };
     } else {
-    setting.sync.services.door43 = setting?.sync?.services?.door43 ? setting?.sync?.services?.door43 : [];
+      setting.sync.services.door43 = setting?.sync?.services?.door43 ? setting?.sync?.services?.door43 : [];
+    }
   }
-}
   setting.project[data.type.flavorType.flavor.name] = data.project[data.type.flavorType.flavor.name];
   logger.debug('updateAgSettings.js', `Updating the ${environment.PROJECT_SETTING_FILE}`);
   await fs.writeFileSync(folder, JSON.stringify(setting));
 };
+
+export const updateWebAgSettings = async (username, projectName, data) => {
+  const result = Object.keys(data.ingredients).filter((key) => key.includes(environment.PROJECT_SETTING_FILE));
+  const folder = `${newPath}/${username}/projects/${projectName}/${result[0]}`;
+  const { data: settings } = await supabaseStorage().download(folder);
+  let setting = {};
+  setting = JSON.parse(await settings.text());
+  if (settings.version !== environment.AG_SETTING_VERSION) {
+    setting.version = environment.AG_SETTING_VERSION;
+    if (!setting.sync && !setting.sync?.services) {
+      setting.sync = { services: { door43: [] } };
+    } else {
+      setting.sync.services.door43 = setting?.sync?.services?.door43 ? setting?.sync?.services?.door43 : [];
+    }
+  }
+  setting.project[data.type.flavorType.flavor.name] = data.project[data.type.flavorType.flavor.name];
+  await supabaseStorage().upload(folder, JSON.stringify(setting), {
+    // cacheControl: '3600',
+    upsert: true,
+  });
+};
+
 export const saveReferenceResource = () => {
   logger.debug('updateAgSettings.js', 'In saveReferenceResource for saving the reference data');
   localforage.getItem('currentProject').then(async (projectName) => {
@@ -38,7 +62,11 @@ export const saveReferenceResource = () => {
               const id = Object.keys(resources.identification.primary[packageInfo.name]);
               if (id[0] === _projectname[1]) {
                 localforage.getItem('userProfile').then(async (value) => {
-                  await updateAgSettings(value?.username, projectName, resources);
+                  if (isElectron()) {
+                    await updateAgSettings(value?.username, projectName, resources);
+                  } else {
+                    await updateWebAgSettings(value?.user?.email, projectName, resources);
+                  }
                 });
               }
             },
