@@ -55,24 +55,24 @@ export default function ProjectsLayout(props) {
         <TopMenuBar />
 
         {title && (
-        <header className="bg-white shadow">
-          {!isTwoCol
-            ? (
-              <div className="mx-auto py-4 px-4 sm:px-4 lg:px-6 border-primary border-b-4 flex items-center justify-between">
-                <div className="flex items-center">
+          <header className="bg-white shadow">
+            {!isTwoCol
+              ? (
+                <div className="mx-auto py-4 px-4 sm:px-4 lg:px-6 border-primary border-b-4 flex items-center justify-between">
                   <div className="flex items-center">
-                    <h1 aria-label="projects" className="text-xl font-bold text-gray-900 uppercase tracking-wider">{showArchived ? 'Archived Projects' : title}</h1>
-                    {header}
+                    <div className="flex items-center">
+                      <h1 aria-label="projects" className="text-xl font-bold text-gray-900 uppercase tracking-wider">{showArchived ? 'Archived Projects' : title}</h1>
+                      {header}
+                    </div>
                   </div>
-                </div>
-                <div className="ml-auto flex">
-                  {isImport && !showArchived
-                    && (
-                      <>
-                        <button
-                          aria-label="import"
-                          type="button"
-                          className="flex text-white ml-5 font-bold text-xs px-3 py-2 rounded-full
+                  <div className="ml-auto flex">
+                    {isImport && !showArchived
+                      && (
+                        <>
+                          <button
+                            aria-label="import"
+                            type="button"
+                            className="flex text-white ml-5 font-bold text-xs px-3 py-2 rounded-full
                                     leading-3 tracking-wider uppercase bg-primary items-center"
                           onClick={openImportPopUp}
                         >
@@ -96,41 +96,41 @@ export default function ProjectsLayout(props) {
                       <button
                         className={`flex text-white ml-5 font-bold text-xs px-3 py-2 rounded-full
                                     leading-3 tracking-wider uppercase ${showArchived ? 'bg-primary' : 'bg-red-600'} items-center`}
-                        type="button"
-                        onClick={toggleArchive}
-                      >
+                          type="button"
+                          onClick={toggleArchive}
+                        >
 
-                        {showArchived ? (
-                          <>
-                            <ComputerDesktopIcon className="h-4 mr-2 text-white" />
-                            <span>Active</span>
-                          </>
-                        ) : (
-                          <>
-                            <ArchiveBoxIcon className="h-4 mr-2 text-white" />
-                            <span>Archived</span>
-                          </>
-                        )}
-                      </button>
+                          {showArchived ? (
+                            <>
+                              <ComputerDesktopIcon className="h-4 mr-2 text-white" />
+                              <span>Active</span>
+                            </>
+                          ) : (
+                            <>
+                              <ArchiveBoxIcon className="h-4 mr-2 text-white" />
+                              <span>Archived</span>
+                            </>
+                          )}
+                        </button>
 
-                    </div>
-                  )}
+                      </div>
+                    )}
+                  </div>
                 </div>
-              </div>
-            )
-            : (
-              <div className="mx-auto px-4 sm:px-4 lg:px-6 border-primary border-b-4 grid grid-cols-2 gap-2">
-                <div className="flex flex-row py-4 items-center">
-                  <h1 className="text-xl font-bold text-gray-900 uppercase tracking-wider">{title}</h1>
-                  {colOne}
+              )
+              : (
+                <div className="mx-auto px-4 sm:px-4 lg:px-6 border-primary border-b-4 grid grid-cols-2 gap-2">
+                  <div className="flex flex-row py-4 items-center">
+                    <h1 className="text-xl font-bold text-gray-900 uppercase tracking-wider">{title}</h1>
+                    {colOne}
+                  </div>
+                  <div className="flex items-end">
+                    {colTwo}
+                  </div>
                 </div>
-                <div className="flex items-end">
-                  {colTwo}
-                </div>
-              </div>
-            )}
+              )}
 
-        </header>
+          </header>
         )}
 
         {children}
