@@ -71,6 +71,7 @@ export default function ProjectList() {
 
   const handleSelectProject = (event, projectName, projectId) => {
     logger.debug('ProjectList.js', 'In handleSelectProject');
+    console.log({ projectName });
     setSelectedProject(projectName);
     localforage.setItem('currentProject', `${projectName}_${projectId}`);
     router.push('/home');
@@ -116,7 +117,7 @@ export default function ProjectList() {
     setCallEditProject(false);
     await FetchProjects();
   };
-// checking if isArchived is true show projects in archive tab else project tab
+  // checking if isArchived is true show projects in archive tab else project tab
   function filterArchive(project) {
     if (project.isArchived === showArchived) {
       return true;
@@ -125,7 +126,6 @@ export default function ProjectList() {
     }
     return false;
   }
-
   return (
     <>
       {callEditProject === false
@@ -470,7 +470,7 @@ export default function ProjectList() {
                               )}
                             </tbody>
                           </table>
-                          {(!starredrow || !unstarredrow) && <div><LoadingScreen /></div> }
+                          {(!starredrow || !unstarredrow) && <div><LoadingScreen /></div>}
                         </div>
                       </div>
                     </div>
@@ -490,7 +490,7 @@ export default function ProjectList() {
               </ProjectContextProvider>
             </AutographaContextProvider>
           </AuthenticationContextProvider>
-)}
+        )}
     </>
   );
 }
